requirements.txt
python-telegram-bot[job-queue]==21.6
openai==1.51.0
